@@ -1,37 +1,40 @@
 {
   "name": "light-bootstrap-dashboard-react",
-  "version": "2.0.0",
+  "version": "2.0.1",
   "dependencies": {
     "@fortawesome/fontawesome-free": "5.15.2",
     "bootstrap": "4.6.0",
-    "chartist": "0.10.1",
+    "chartist": "0.11.4",
     "gulp": "4.0.2",
-    "gulp-append-prepend": "1.0.8",
-    "node-sass": "4.14.1",
-    "react": "17.0.1",
-    "react-bootstrap": "1.4.3",
-    "react-chartist": "0.14.3",
-    "react-dom": "17.0.1",
+    "gulp-append-prepend": "1.0.9",
+    "react": "18.2.0",
+    "reactstrap": "8.8.1",
+    "react-bootstrap": "2.4.0",
+    "react-chartist": "0.14.4",
+    "react-dom": "18.2.0",
     "react-notification-alert": "0.0.13",
     "react-router": "5.2.0",
     "react-router-dom": "5.2.0",
-    "react-scripts": "4.0.1"
+    "react-scripts": "5.0.1",
+    "sass": "1.53.0"
   },
   "optionalDependencies": {
     "jquery": "3.5.1",
-    "reactstrap": "8.9.0",
-    "typescript": "4.1.3"
+    "typescript": "4.7.4"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build && gulp licenses",
-    "test": "react-scripts test --env=jsdom",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject",
     "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && npm start",
-    "compile:scss": "node-sass src/assets/scss/light-bootstrap-dashboard-react.scss src/assets/css/light-bootstrap-dashboard-react.css",
-    "minify:scss": "node-sass src/assets/scss/light-bootstrap-dashboard-react.scss src/assets/css/light-bootstrap-dashboard-react.min.css --output-style compressed",
-    "map:scss": "node-sass src/assets/scss/light-bootstrap-dashboard-react.scss src/assets/css/light-bootstrap-dashboard-react.css --source-map true",
-    "build:scss": "npm run compile:scss && npm run minify:scss && npm run map:scss"
+    "compile:scss": "sass src/assets/scss/light-bootstrap-dashboard-pro-react.scss src/assets/css/light-bootstrap-dashboard-pro-react.css",
+    "minify:scss": "sass src/assets/scss/light-bootstrap-dashboard-pro-react.scss src/assets/css/light-bootstrap-dashboard-pro-react.min.css --style compressed",
+    "build:scss": "npm run compile:scss && npm run minify:scss"
+  },
+  "peerDependencies": {
+    "react": ">=16.8.0",
+    "react-dom": ">=16.8.0"
   },
   "browserslist": {
     "production": [
