@@ -9,7 +9,7 @@ gulp.task("licenses", async function () {
       gap.prependText(`/*!
 
 =========================================================
-* Light Bootstrap Dashboard React - v2.0.0
+* Light Bootstrap Dashboard React - v2.0.1
 =========================================================
 
 * Product Page: https://www.creative-tim.com/product/light-bootstrap-dashboard-react
@@ -32,7 +32,7 @@ gulp.task("licenses", async function () {
       gap.prependText(`<!--
 
 =========================================================
-* Light Bootstrap Dashboard React - v2.0.0
+* Light Bootstrap Dashboard React - v2.0.1
 =========================================================
 
 * Product Page: https://www.creative-tim.com/product/light-bootstrap-dashboard-react
@@ -55,7 +55,7 @@ gulp.task("licenses", async function () {
       gap.prependText(`/*!
 
 =========================================================
-* Light Bootstrap Dashboard React - v2.0.0
+* Light Bootstrap Dashboard React - v2.0.1
 =========================================================
 
 * Product Page: https://www.creative-tim.com/product/light-bootstrap-dashboard-react
