@@ -1,11 +1,11 @@
 /*!
 
 =========================================================
-* Light Bootstrap Dashboard React - v2.0.0
+* Light Bootstrap Dashboard React - v2.0.1
 =========================================================
 
 * Product Page: https://www.creative-tim.com/product/light-bootstrap-dashboard-react
-* Copyright 2020 Creative Tim (https://www.creative-tim.com)
+* Copyright 2022 Creative Tim (https://www.creative-tim.com)
 * Licensed under MIT (https://github.com/creativetimofficial/light-bootstrap-dashboard-react/blob/master/LICENSE.md)
 
 * Coded by Creative Tim
@@ -31,7 +31,7 @@ function FixedPlugin({
   color,
   setColor,
   image,
-  setImage,
+  setImage
 }) {
   // constructor(props) {
   //   super(props);
