@@ -10,7 +10,7 @@ import {
   Nav,
   Container,
   Row,
-  Col,
+  Col
 } from "react-bootstrap";
 
 function User() {
@@ -156,10 +156,7 @@ function User() {
               <div className="card-image">
                 <img
                   alt="..."
-                  src={
-                    require("assets/img/photo-1431578500526-4d9613015464.jpeg")
-                      .default
-                  }
+                  src={require("assets/img/photo-1431578500526-4d9613015464.jpeg")}
                 ></img>
               </div>
               <Card.Body>
@@ -168,7 +165,7 @@ function User() {
                     <img
                       alt="..."
                       className="avatar border-gray"
-                      src={require("assets/img/faces/face-3.jpg").default}
+                      src={require("assets/img/faces/face-3.jpg")}
                     ></img>
                     <h5 className="title">Mike Andrew</h5>
                   </a>
